refactor: 리다이렉트 페이지 VAC 패턴 적용
